+++ net/http/roundtrip_test.go
@@ -1,4 +1,4 @@
-package roundtrip
+package http
 
 import (
 	"net/http"
